+++ packages/channel/.eslintrc.cjs
@@ -3,6 +3,6 @@ module.exports = {
   extends: ["@repo/eslint-config/library.cjs"],
   parser: "@typescript-eslint/parser",
   parserOptions: {
-    project: true,
+    project: './eslint.tsconfig.json',
   },
 };
