+++ packages/channel/eslint.tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  },
+  "extends": "@repo/typescript-config/app.json",
+  "include": ["./src/**/*", "./test/**/*.test.ts"]
+}
