+++ packages/redis/src/ioredis.ts
@@ -77,3 +77,7 @@ export const redisDelPattern = (pattern: string): void => {
     }
   });
 };
+
+export const redisQuit = async () => {
+  await ioredis.quit();
+};
